+++ src/lib/export/index.ts
@@ -20,3 +20,5 @@ export * from "./ppfCsv";
 export * from "./ppfJson";
 export * from "./sipCsv";
 export * from "./sipJson";
+export * from "./ssyCsv";
+export * from "./ssyJson";